# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sufrix API",
    "description": "Sufrix POS — multi-tenant cafe and restaurant management. The Rust backend is the source of truth for this spec; the Flutter teller app and React management dashboard consume the generated openapi.json.",
    "contact": {
      "name": "Sufrix",
      "url": "https://sufrix.app"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    },
    {
      "url": "https://api.sufrix.app",
      "description": "Production"
    }
  ],
  "paths": {
    "/branches": {
      "get": {
        "tags": [
          "branches"
        ],
        "operationId": "list_branches",
        "parameters": [
          {
            "name": "org_id",
            "in": "query",
            "description": "Organization whose branches to list. Must match the caller's JWT org.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of branches in the organization",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request — validation failed or malformed input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — insufficient permission or wrong org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict — FK or domain invariant violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_jwt": []
          }
        ]
      },
      "post": {
        "tags": [
          "branches"
        ],
        "operationId": "create_branch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Branch created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "description": "Bad request — validation failed or malformed input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — insufficient permission or wrong org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict — FK or domain invariant violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_jwt": []
          }
        ]
      }
    },
    "/branches/{id}": {
      "get": {
        "tags": [
          "branches"
        ],
        "operationId": "get_branch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Branch ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested branch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "description": "Bad request — validation failed or malformed input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — insufficient permission or wrong org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict — FK or domain invariant violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_jwt": []
          }
        ]
      },
      "put": {
        "tags": [
          "branches"
        ],
        "operationId": "update_branch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Branch ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Branch updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "description": "Bad request — validation failed or malformed input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — insufficient permission or wrong org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict — FK or domain invariant violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_jwt": []
          }
        ]
      },
      "delete": {
        "tags": [
          "branches"
        ],
        "operationId": "delete_branch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Branch ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Branch deleted (soft delete)"
          },
          "400": {
            "description": "Bad request — validation failed or malformed input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — insufficient permission or wrong org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict — FK or domain invariant violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_jwt": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Branch": {
        "type": "object",
        "required": [
          "id",
          "org_id",
          "name",
          "timezone",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "address": {
            "type": [
              "string",
              "null"
            ],
            "example": "26 July Corridor, Zamalek, Cairo"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_active": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "example": "Zamalek"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "org_logo_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Convenience field — populated from the parent org's `logo_url`."
          },
          "phone": {
            "type": [
              "string",
              "null"
            ],
            "example": "+201234567890"
          },
          "printer_brand": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrinterBrand"
              }
            ]
          },
          "printer_ip": {
            "type": [
              "string",
              "null"
            ],
            "example": "192.168.1.50"
          },
          "printer_port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "example": 9100
          },
          "timezone": {
            "type": "string",
            "description": "IANA timezone name. Defaults to `Africa/Cairo`.",
            "example": "Africa/Cairo"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateBranchRequest": {
        "type": "object",
        "required": [
          "org_id",
          "name"
        ],
        "properties": {
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string",
            "example": "Zamalek"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "printer_brand": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrinterBrand"
              }
            ]
          },
          "printer_ip": {
            "type": [
              "string",
              "null"
            ]
          },
          "printer_port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "TCP port for the receipt printer. Defaults to `9100` if absent.",
            "example": 9100
          },
          "timezone": {
            "type": [
              "string",
              "null"
            ],
            "description": "IANA timezone name. Defaults to `Africa/Cairo` if absent.",
            "example": "Africa/Cairo"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Wire shape of every error JSON. Keep in lockstep with\n`AppError::error_response` below.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message.",
            "example": "Branch not found"
          }
        }
      },
...
      "PrinterBrand": {
        "type": "string",
        "enum": [
          "star",
          "epson"
        ]
      },
      "UpdateBranchRequest": {
        "type": "object",
        "description": "PATCH-style update. Fields fall into three categories:\n\n- **Absent** from JSON → keep existing value.\n- **Present as `null`** (only the `printer_*` fields) → clear the column.\n- **Present as a value** → set to that value.\n\nOpenAPI cannot express the absent-vs-null distinction cleanly, so all\nfields are documented as optional and nullable. Clients targeting this\nendpoint should send only the fields they want to change.",
        "properties": {
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "printer_brand": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrinterBrand"
              }
            ]
          },
          "printer_ip": {
            "type": [
              "string",
              "null"
            ]
          },
          "printer_port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "timezone": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from `POST /auth/login` (email/password) or `POST /auth/pin-login` (teller PIN). Send as `Authorization: Bearer <token>`."
      }
    }
  },
...
}